perf: js: load username autocompletion asynchronously and on-demand
